+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Ruby 4.0 support (no changes required)
 * Support procs on validators for minlength/maxlength, and improve validators logic across the board to match Rails [#1859](https://github.com/heartcombo/simple_form/pull/1859)
+* Fix label `for` attribute not matching input `id` when collection contains a nil value [#1840](https://github.com/heartcombo/simple_form/issues/1840)
 
 ## 5.4.0